Give each experiment its own database, and update every database an image belongs to
Change:
  - Creating an experiment creates its database, named after it. The
    experiment dialog lists every experiment in every database, marked
    with where it is stored; experiments already in the standard
    database stay there and can be moved to a database of their own,
    one at a time, on request.
  - Removing an experiment deletes its database, after a warning. An
    experiment still stored in the standard database is removed from it
    and its images and results stay; the database in use cannot be
    removed.
  - Adding an image to an experiment brings it into that experiment's
    database. If the standard database already holds its data, the user
    chooses whether to migrate it there or copy it.
  - An analysis is written to the standard database and to every
    experiment database the image belongs to, so re-analysing an image
    updates each experiment that uses it. Before a single or a batch
    analysis starts, the user is told how many experiment databases will
    be updated and can decline, which stops the analysis. A batch asks
    once.
  - The export, the statistics and the main window's experiment view
    read an experiment from its own database, whichever one is in use.
  - The settings dialog's main tab has a list of every available
    database; choosing one and pressing OK switches the program to it.
    The image list follows the switch; two of its helpers read the
    standard database whatever was in use.
  - Moving one experiment out of a database that holds others no longer
    takes the other experiments' memberships, or the data they still
    need, with it; copying an image into an experiment no longer copies
    its memberships of other experiments.
  - Channel names and experiment memberships are stored as text. Their
    columns were declared as integers, so a value that looked like a
    number was stored as one: an experiment named 2024 came back as the
    number 2024. This is schema version 5; existing databases are
    converted from the settings dialog.
  - The nucleus watershed is flooded with a true negation of the
    distance map. The map is stored as unsigned bytes, and negating it
    wrapped zero to the lowest value instead of the highest. No real
    image measured was affected.

AI assistance: The changes in this commit were produced with assistance
from Anthropic's Claude models:
  - Claude Opus 5.5 (claude-opus-5-5) - implementation, the comparison
    of the old and new watershed, and the verification harnesses

Review: Every change in this commit was reviewed and approved manually
by RW. No change was applied without human inspection.

+++ core/detector_modules/NucleusMapper.py
@@ -269,6 +269,17 @@ class NucleusMapper(AreaMapper):
         :param line: Toggle to draw a line between segmented areas
         :return: The segmented binary map
         """
-        # Create watershed segmentation based on centers
-        return watershed(-edm, cmask, mask=mask, watershed_line=line)
+        # Create watershed segmentation based on centers.
+        #
+        # NEGATED AS FLOAT, since 2026-09-24. The map arrives as uint8, and `-edm` on an unsigned
+        # array wraps modulo 256: 1..255 become 255..1, which keeps their order, but 0 stays 0 --
+        # the LOWEST value where a true negation makes it the highest. A foreground pixel quantises
+        # to 0 once the distance range exceeds ~383 px (a threshold that caught the whole frame),
+        # and those pixels were flooded first. The upcast costs nothing: watershed converts its
+        # input to float64 anyway, but only AFTER the minus had wrapped.
+        #
+        # Output-identical on every image where no foreground pixel quantises to 0, which is every
+        # real image measured -- the flood order depends only on the ordering of the values, and
+        # 256 - x orders 1..255 exactly as -x does
+        return watershed(-edm.astype(np.float64), cmask, mask=mask, watershed_line=line)
